+++ app/page.module.scss
@@ -1,4 +1,4 @@
 .main {
-    background-color: #000000;
-    height: 100vh;
+    background-color: #121212;
+    height: 100svh;
 }
