--- components/Footer/Footer.module.scss
@@ -1,152 +1,0 @@
-.footerContainer {
-    display: flex;
-    justify-content: center;
-    width: 100%;
-    padding: 0 10%;
-    background-color: #000000;
-
-    @media screen and (min-width: 950px) {
-        transform: scale(0.8) translateY(-30%);
-    }
-}
-
-.footer {
-    display: flex;
-    flex-direction: column;
-    row-gap: 2.5rem;
-    width: 100%;
-    background-color: #000000;
-
-    @media screen and (max-width: 900px) {
-        padding-bottom: 7.5rem;
-    }
-
-    @media screen and (max-width: 700px) {
-        padding-bottom: 4rem;
-    }
-
-    > div:nth-child(1) {
-        display: flex;
-        flex-flow: column nowrap;
-        align-items: center;
-        justify-content: center;
-        align-self: center;
-        max-width: max(90%, calc(100vw / (1920 / 1220)));
-    }
-}
-
-.title {
-    display: block;
-    font-size: 2.5rem;
-    @media screen and (max-width: 1280px) {
-        font-size: 2.5rem;
-    }
-
-    @media screen and (max-width: 1024px) {
-        font-size: 2rem;
-    }
-
-    @media screen and (max-width: 900px) {
-        font-size: 1.75rem;
-    }
-
-    @media screen and (max-width: 700px) {
-        font-size: 1.25rem;
-        //white-space: nowrap;
-    }
-    font-weight: 600;
-    color: #ffffff;
-    text-align: center;
-}
-
-.icons {
-    display: flex;
-    flex-wrap: wrap;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    margin-top: 2rem;
-    gap: 4rem;
-
-    @media screen and (max-width: 900px) {
-        margin-top: 1.5rem;
-        gap: 2.5rem;
-    }
-
-    @media screen and (max-width: 700px) {
-        margin-top: 1.5rem;
-        column-gap: 20%;
-        row-gap: 2rem;
-    }
-
-    .icon {
-        display: inline-flex;
-        justify-content: center;
-        align-items: center;
-        background-color: #305e42;
-        // padding: min(1vw, 1rem);
-        border-radius: 9999px;
-        width: 4rem;
-        height: 4rem;
-
-        @media screen and (max-width: 1280px) {
-            width: 4rem;
-            height: 4rem;
-        }
-
-        @media screen and (max-width: 1024px) {
-            width: 3.5rem;
-            height: 3.5rem;
-        }
-
-        @media screen and (max-width: 900px) {
-            width: 3rem;
-            height: 3rem;
-        }
-
-        @media screen and (max-width: 600px) {
-            width: 2rem;
-            height: 2rem;
-        }
-
-        &:hover {
-            cursor: pointer;
-            transform: scale(1.05);
-        }
-
-        .image {
-            width: 50%;
-            height: 50%;
-        }
-    }
-}
-
-.newsletterTitle {
-    display: block;
-    font-size: 2rem;
-    margin: 2.5rem 0 3rem;
-
-    @media screen and (max-width: 1280px) {
-        font-size: 2rem;
-        margin: 2.5rem 0 2.5rem;
-    }
-
-    @media screen and (max-width: 1024px) {
-        font-size: 1.5rem;
-        margin: 2rem 0 2.5rem;
-    }
-
-    @media screen and (max-width: 900px) {
-        font-size: 1.25rem;
-        margin: 1rem 0 1.5rem;
-    }
-
-    @media screen and (max-width: 700px) {
-        font-size: 1.125rem;
-        margin: 1rem 0 1.5rem;
-    }
-    font-weight: 600;
-    color: #ffffff;
-    text-align: center;
-    // white-space: nowrap;
-}
