+++ components/Hero/styles.module.scss
@@ -1,27 +1,39 @@
 @use "/app/colors" as *;
 
-.heroMain {
+.main {
     width: 100%;
-    height: 65vh;
-    background-color: #000000;
-    overflow-x: clip;
+    height: 100%;
+    overflow: clip;
+    padding: 0 0.25rem;
 
-    @media screen and (max-width: 800px) {
-        height: 50vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    gap: 1.5rem;
+
+    @media screen and (max-width: 500px) {
+        gap: 1rem;
     }
 }
 
-.heroContainer {
-    width: 100%;
-    height: calc(100% - 175px);
-    position: relative;
-    background-repeat: no-repeat;
-    background-size: cover;
-    background-position: center;
+.logoAnimation {
+    height: 40svh;
+    width: 40svh;
 
-    @media screen and (max-width: 800px) {
-        height: calc(100% - 150px);
+    @media screen and (max-height: 1200px) {
+        height: 30svh;
+        width:  30svh;
     }
+
+    @media screen and (max-height: 800px) {
+        height: 15svh;
+        width:  15svh;
+    }
+}
+
+.title {
+    width: 100%;
 
     h1 {
         font-weight: 700;
@@ -50,53 +62,17 @@
             font-size: 35px;
         }
     }
-
-    .topHeader {
-        position: absolute;
-        width: 100%;
-        z-index: 2;
-        top: 95%;
-
-        @media screen and (max-width: 900px) {
-            top: 95%;
-        }
-
-        @media screen and (max-height: 500px) {
-            top: 95%;
-        }
-    }
-
-    .logoAnimation {
-        position: absolute;
-        left: 50%;
-        transform: translate(-50%, 0);
-        top: 0;
-        height: 95%;
-        aspect-ratio: 1;
-
-        @media screen and (max-height: 500px) {
-            height: 95%;
-        }
-    }
 }
 
-.heroFooter {
-    position: absolute;
-    background-color: #000000;
-    line-height: 50px;
+.subtitle {
     width: 100%;
-    overflow-y: hidden;
-    margin-top: 35px;
-
     display: flex;
     flex-direction: column;
     align-items: center;
 
-    .subtitle {
-        margin-top: 30px;
+    .text {
         font-weight: 500;
         font-size: 45px;
-        height: 60px;
         color: #fcdc8f;
         text-shadow: 0px 1px 20px rgba(252, 220, 143, 0.76);
         animation-name: glow;
@@ -105,7 +81,6 @@
         animation-direction: alternate;
 
         @media screen and (max-width: 800px) {
-            margin-top: 0;
             font-size: 32px;
         }
 
@@ -124,6 +99,118 @@
                 text-shadow:
                     0px 0px 20px #fff,
                     0px 0px 4px #bd7c62;
+            }
+        }
+    }
+}
+
+.newsletter {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 1rem;
+
+    .newsletterTitle {
+        font-size: 2.5rem;
+
+        @media screen and (max-width: 1280px) {
+            font-size: 2.5rem;
+        }
+
+        @media screen and (max-width: 1024px) {
+            font-size: 2rem;
+        }
+
+        @media screen and (max-width: 900px) {
+            font-size: 1.75rem;
+        }
+
+        @media screen and (max-width: 700px) {
+            font-size: 1.25rem;
+        }
+        font-weight: 600;
+        color: #ffffff;
+        text-align: center;
+    }
+}
+
+.socials {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 1rem;
+
+    .title {
+        font-size: 2.5rem;
+        @media screen and (max-width: 1280px) {
+            font-size: 2.5rem;
+        }
+
+        @media screen and (max-width: 1024px) {
+            font-size: 2rem;
+        }
+
+        @media screen and (max-width: 900px) {
+            font-size: 1.75rem;
+        }
+
+        @media screen and (max-width: 700px) {
+            font-size: 1.25rem;
+        }
+        font-weight: 600;
+        color: #ffffff;
+        text-align: center;
+    }
+
+    .icons {
+        display: flex;
+        flex-wrap: wrap;
+        flex-direction: row;
+        justify-content: center;
+        align-items: center;
+        gap: 4rem;
+
+        @media screen and (max-width: 900px) {
+            gap: 2.5rem;
+        }
+
+        @media screen and (max-width: 700px) {
+            column-gap: 20%;
+            row-gap: 1.5rem;
+        }
+
+        .icon {
+            display: inline-flex;
+            justify-content: center;
+            align-items: center;
+            background-color: #305e42;
+            border-radius: 9999px;
+            width: 4rem;
+            height: 4rem;
+
+            @media screen and (max-width: 1280px) {
+                width: 4rem;
+                height: 4rem;
+            }
+
+            @media screen and (max-width: 1024px) {
+                width: 3.5rem;
+                height: 3.5rem;
+            }
+
+            @media screen and (max-width: 900px) {
+                width: 3rem;
+                height: 3rem;
+            }
+
+            &:hover {
+                cursor: pointer;
+                transform: scale(1.05);
+            }
+
+            .image {
+                width: 50%;
+                height: 50%;
             }
         }
     }
